Create local bucketing interface for flexibility
