+++ DevCycle.SDK.Server.Local.Benchmark/BenchmarkTests.cs
@@ -40,7 +40,7 @@ namespace DevCycle.SDK.Server.Local.Benchmark
             mockHttp.When("https://events*")
                 .Respond(HttpStatusCode.Created, "application/json",
                     "{}");
-            var localBucketing = new LocalBucketing();
+            var localBucketing = new WASMLocalBucketing();
             var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
             var sdkKey = $"dvc_server_{Guid.NewGuid().ToString().Replace('-','_')}_hash";
             localBucketing.StoreConfig(sdkKey, config);
