+++ DevCycle.SDK.Server.Local.MSTests/Mocks/DevCycleClient.cs
@@ -24,7 +24,7 @@ public class DevCycleTestClient
         mockHttp.When("https://events*")
             .Respond(HttpStatusCode.Created, "application/json",
                 "{}");
-        var localBucketing = new LocalBucketing();
+        var localBucketing = new WASMLocalBucketing();
         var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
         var sdkKey = $"dvc_server_{Guid.NewGuid().ToString().Replace('-', '_')}_hash";
         localBucketing.StoreConfig(sdkKey, config);
